Migrate `impl AuthQuery for BaseCoinApp` to `app/modules/auth.rs`
